allball_qa_insight_poc/requirements.txt
altair==5.3.0
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.2.0
backports.tarfile==1.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
decorator==5.1.1
docutils==0.21.2
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==7.2.1
ipython==8.26.0
ipywidgets==8.1.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.11
keyring==25.2.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
nh3==0.2.18
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyloco==0.0.139
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
readme_renderer==44.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.19.0
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.11.0
snowflake-snowpark-python==1.19.0
sortedcontainers==2.4.0
st-circular-progress==0.3.9
stack-data==0.6.3
streamlit==1.31.1
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.0
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
twine==5.1.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
ushlex==0.99.1
validators==0.32.0
watchdog==4.0.1
wcwidth==0.2.13
websocket-client==1.8.0
widgetsnbextension==4.0.11
zipp==3.19.2
